% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms_helpers.R
\name{convert_bitmaps_to_counts}
\alias{convert_bitmaps_to_counts}
\title{Performs R.BITCOUNT operation on bitmaps to produce counts}
\usage{
convert_bitmaps_to_counts(bitmaps, roaring_client)
}
\arguments{
\item{bitmaps}{a bitmap dataframe}

\item{roaring_client}{a Redis client e.g.:\cr
\code{roaring_client <- redis_client('redis://127.0.0.1:6380/0')}}
}
\value{
\tabular{rrrrr}{
brand \tab usage \tab counts\cr
Pepsi \tab   1  \tab    1\cr
Fanta \tab    1 \tab    2\cr
Coke  \tab   1  \tab    2\cr
}
}
\description{
Performs R.BITCOUNT operation on bitmaps to produce counts
}
