% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms_helpers.R
\name{get_respondent_indexes}
\alias{get_respondent_indexes}
\title{Retrieve respondent indexes for given root_keys}
\usage{
get_respondent_indexes(root_keys, roaring_client)
}
\arguments{
\item{root_keys}{a vector of root keys e.g.: \code{c("redis:bitops:f0f03fec88a8effcd6e131fdc9407232de0387eb")}}

\item{roaring_client}{a Redis client e.g.:\cr
\code{roaring_client <- redis_client('redis://127.0.0.1:6380/0')}}
}
\description{
Retrieve respondent indexes for given root_keys
}
