% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms_helpers.R
\name{member_count}
\alias{member_count}
\title{Generates member counts}
\usage{
member_count(respondent_level_bitmaps, roaring_client)
}
\arguments{
\item{respondent_level_bitmaps}{a respondent level bitmap dataframe}

\item{roaring_client}{a Redis client e.g.:\cr
\code{roaring_client <- redis_client('redis://127.0.0.1:6380/0')}}
}
\value{
\tabular{rrrrr}{
member_count \tab root_key\cr
<int> \tab <chr>\cr
 1 \tab redis:bitops:3gfbiagag29bdeh199hab4482adgcbd94h97847f\cr
 2 \tab redis:bitops:0hb9f0ciii1cbcb7che274b36792a82h5gef1agf\cr
}
}
\description{
Generates member counts
}
