% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms_helpers.R
\name{redis_client}
\alias{redis_client}
\title{Create a Redis client}
\usage{
redis_client(url)
}
\arguments{
\item{url}{example: \code{'redis://127.0.0.1:6380/1'}}
}
\value{
a client for connecting to Redis
}
\description{
Create a Redis client
}
