% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms_helpers.R
\name{redis_roaring_client}
\alias{redis_roaring_client}
\title{Create a RedisRoaring client}
\usage{
redis_roaring_client(url)
}
\arguments{
\item{url}{example: \code{'redis://127.0.0.1:6380/1'}}
}
\value{
a client for connecting to RedisRoaring
}
\description{
Create a RedisRoaring client
}
