% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms_helpers.R
\name{to_respondent_level_data}
\alias{to_respondent_level_data}
\title{Transforms bitmaps to respondent level data}
\usage{
to_respondent_level_data(bitmaps, roaring_client)
}
\arguments{
\item{bitmaps}{a Ruby array of hashes}

\item{roaring_client}{RedisRoaring client e.g.: \cr
\code{confl <- redis_config(url='redis://127.0.0.1:6380/1')} \cr
\code{roaring_client <- redux::hiredis(confl)}}
}
\value{
\tabular{rrrrr}{
 brand \tab usage             \tab                                       root_key  \tab respondent_id\cr
 Pepsi \tab    1 response_id|ms8b129767|cds1c52f958|s9644a47e|cdsd0eb5e70|i1 \tab            2\cr
 Fanta \tab    1 response_id|ms8b129767|cds1c52f958|sa1a4c184|cdsd0eb5e70|i1 \tab           3\cr
 Fanta \tab    1 response_id|ms8b129767|cds1c52f958|sa1a4c184|cdsd0eb5e70|i1 \tab            4\cr
 Coke  \tab   1 response_id|ms8b129767|cds1c52f958|s50bf8369|cdsd0eb5e70|i1   \tab          1\cr
 Coke  \tab   1 response_id|ms8b129767|cds1c52f958|s50bf8369|cdsd0eb5e70|i1  \tab           2\cr
}
}
\description{
Transforms bitmaps to respondent level data
}
