% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms_helpers.R
\name{unique_key}
\alias{unique_key}
\title{Generates a random unique key. Intended usage is for the creation of temporary bitmaps}
\usage{
unique_key()
}
\value{
"redis:bitops:dbhhg8i2hga7ee7g6bh5hjhdd5217bjh075jh94c"
}
\description{
Generates a random unique key. Intended usage is for the creation of temporary bitmaps
}
