% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms_helpers.R
\name{write_respondent_indexes}
\alias{write_respondent_indexes}
\title{Writes given indexes to RedisRoaring under a uniquely generated key}
\usage{
write_respondent_indexes(indexes, roaring_client)
}
\arguments{
\item{indexes}{character variable e.g.: \code{c("1, 3, 4", "2")}}

\item{roaring_client}{a Redis client e.g.:\cr
\code{roaring_client <- redis_client('redis://127.0.0.1:6380/0')}}
}
\value{
"redis:bitops:8dd1b989996gg8523633c06h7d4026067fc1cfc6" "redis:bitops:06d01cfc2gijj2da3hjdc36271ga6g15bcbc0773"
}
\description{
Writes given indexes to RedisRoaring under a uniquely generated key
}
