% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms_helpers.R
\name{to_respondent_level_data}
\alias{to_respondent_level_data}
\title{Transforms bitmaps to respondent level data}
\usage{
to_respondent_level_data(bitmaps, roaring_client)
}
\arguments{
\item{bitmaps}{a bitmap dataframe}

\item{roaring_client}{a Redis client e.g.:\cr
\code{roaring_client <- redis_client('redis://127.0.0.1:6380/0')}}
}
\value{
\tabular{rrrrr}{
 brand \tab usage\tab               respondent_id\cr
 Pepsi \tab    1 \tab               2\cr
 Fanta \tab    1 \tab               3\cr
 Fanta \tab    1 \tab               4\cr
 Coke  \tab    1 \tab               1\cr
 Coke  \tab    1 \tab               2\cr
}
}
\description{
Transforms bitmaps to respondent level data
}
